Tighten the 'with couchdb mocked out' assertion to correctly report errors.
The previous formulation ended up swallowing the exception from the nested
assertion - the most you'd happent to see when something was wrong was an
error thrown trying to close() something already closed. While this was
enough to indicate an error, it is completely the wrong output and actually
indicative of a separate problem.

First ensure cleanUp can only be called once. Then wrap nested assertions
such that we call cleanUp then resolve or reject on their success/failure.